General refactoring
