+++ RevitTranslator/Models/DocumentTranslationEntityGroup.cs
@@ -2,6 +2,6 @@ namespace RevitTranslator.Models;
 
 public sealed class DocumentTranslationEntityGroup(Document document)
 {
-    public Document Document { get; init; } = document;
+    public Document Document { get; } = document;
     public List<TranslationEntity> TranslationEntities { get; } = [];
 }
